+++ Evaluator.Modulators.cs
@@ -22,7 +22,7 @@
 
 /*****************************************************************************
  *                                                                           *
- *  SolusEngine.cs                                                           *
+ *  Evaluator.cs                                                           *
  *                                                                           *
  *  The central core of processing in Solus. Does some rudimentary parsing   *
  *    and evaluation and stuff.                                              *
@@ -36,7 +36,7 @@ using System.Diagnostics;
 
 namespace MetaphysicsIndustries.Solus
 {
-    //public partial class SolusEngine
+    //public partial class Evaluator
     //{
     //    public static float Convert24gToFloat(float value)
     //    {
